added drawing functionality
uses event listeners to track the position of the mouse when it
is pressed down. Position is tracked relative to the canvas.
It works, but not very well. The pen is too small, and the marks
are sporadic. There is also an offset from the cursor, implying that
the coordinates are not being calculated correctly.